Enum osmia::model::stmt::stmt::Stmt

pub enum Stmt {

    Block(Block),
    Raw(String),
    NonPrintable(String),
    NewLine,
    NewLineNonPrintable,
    Expr(Expr),
    Print(Print),
    Comment(String),
    Assign(Assign),
    If(If),
    While(ConditionalStmt),
    For(For),
    Break,
    Continue,
    Return(Return),
    Function(Function),

}

Statements

A statement is a piece of code that can be executed.

Statements that begin with a special keyword (such as if, for, while, etc.), must have the keyword directly after the delimiter, without any space in between.

code / block

A block is a group of statements.

use osmia::Osmia;

let mut osmia = Osmia::default();
assert_eq!(osmia.run_code(r#"
Hello world!
1 + 1 = {{ 1 + 1 }}
{{ "uppercase"?upper() }}
"#.trim()).unwrap(),
"Hello world!\n1 + 1 = 2\nUPPERCASE".to_string());

expr

Allows to output an Expr.

use osmia::Osmia;

let mut osmia = Osmia::default();
assert_eq!(osmia.run_code("{{ math.PI }}").unwrap(), "3.141592653589793".to_string());
assert_eq!(osmia.run_code("{{ 1 + 1 }}").unwrap(), "2".to_string());

print

Allows to print in the stdout the given Expr.

use osmia::Osmia;

let mut osmia = Osmia::default();
osmia.run_code("{{print 1 + 1 }}").unwrap(); // prints "2"

comment

Allows to add a comment to the code. It is ignored while interpreting the code.

If there is an starting delimiter `{{` inside a comment, an ending delimiter `}}` is also expected inside the comment.

use osmia::Osmia;

let mut osmia = Osmia::default();
assert_eq!(osmia.run_code(r#"{{# This is a comment }}"#).unwrap(), "".to_string());
assert_eq!(osmia.run_code(r#"{{#
  This is a multi line comment.
  Code can be added here and it will be ignored as long as there is an ending delimiter
  {{ 1 + 1 }}
}}"#).unwrap(), "".to_string());
assert!(osmia.run_code(r#"{{#
  {{ 1 + 1
}}"#).is_err());

assign

Allows to assign an Expr to a variable in the Context.

use osmia::Osmia;

let mut osmia = Osmia::default();
assert_eq!(osmia.run_code("{{ name = \"Marvin\" }}").unwrap(), "".to_string());
assert_eq!(osmia.run_code("{{ name = \"Marvin\" }}{{ name }}").unwrap(), "Marvin".to_string());
assert_eq!(osmia.run_code("{{ name }}").unwrap(), "Marvin".to_string()); // Stored between calls

if, elseif, else

Executes the Block based on boolean conditionals.

use osmia::Osmia;

let code = r#"
{{if name == "Marvin"}}
	Hello Marvin!
{{elseif name == "R2D2"}}
	You are not Marvin, but you are R2D2!
{{else}}
	You are neither Marvin nor R2D2.
{{fi}}
"#.trim();
let mut osmia = Osmia::try_from(r#"{ "name": "Marvin" }"#).unwrap();
assert_eq!(osmia.run_code(code).unwrap(), "\tHello Marvin!\n".to_string());
let mut osmia = Osmia::try_from(r#"{ "name": "R2D2" }"#).unwrap();
assert_eq!(osmia.run_code(code).unwrap(), "\tYou are not Marvin, but you are R2D2!\n".to_string());
let mut osmia = Osmia::try_from(r#"{ "name": "C3PO" }"#).unwrap();
assert_eq!(osmia.run_code(code).unwrap(), "\tYou are neither Marvin nor R2D2.\n".to_string());

while

Executes the Block while the condition is true.

use osmia::Osmia;

let mut osmia = Osmia::default();
assert_eq!(osmia.run_code(r#"{{i = 0}}{{while i < 10}}{{i}}{{i = i + 1}}{{done}}"#).unwrap(), "0123456789".to_string());

for

Executes the Block for each item in the Expr.

use osmia::Osmia;

let mut osmia = Osmia::default();
assert_eq!(osmia.run_code(r#"{{for i in [1, 2, 3]}}{{i}}{{done}}"#).unwrap(), "123".to_string());

break

Breaks out of the current Block.

use osmia::Osmia;

let code = r#"
{{for i in [1, 2, 3]}}
	{{if i == 2}}
		{{break}}
	{{fi}}
	{{i}}
{{done}}
"#.trim();
let mut osmia = Osmia::default();
assert_eq!(osmia.run_code(code).unwrap(), "\t1\n".to_string());

continue

Continues with next iteration of the current Block.

use osmia::Osmia;

let code = r#"
{{for i in [1, 2, 3]}}
	{{if i == 2}}
		{{continue}}
	{{fi}}
	{{i}}
{{done}}
"#.trim();
let mut osmia = Osmia::default();
assert_eq!(osmia.run_code(code).unwrap(), "\t1\n\t3\n".to_string());

function

Defines a new Function.

use osmia::Osmia;

let code = r#"
{{fn custom_add; a, b}}
	{{return a + b}}
{{done}}
{{custom_add(1, 2)}}
"#.trim();
let mut osmia = Osmia::default();
assert_eq!(osmia.run_code(code).unwrap(), "3".to_string());

return

Returns from the given Expr of the current Function.

use osmia::Osmia;

let code = r#"
{{fn custom_add; a, b}}
	{{return a + b}}
{{done}}
{{custom_add(1, 2)}}
"#.trim();
let mut osmia = Osmia::default();
assert_eq!(osmia.run_code(code).unwrap(), "3".to_string());

Variants

Block(Block)

Raw(String)

NonPrintable(String)

NewLine

NewLineNonPrintable

Expr(Expr)

Print(Print)

Comment(String)

Assign(Assign)

If(If)

While(ConditionalStmt)

For(For)

Break

Continue

Return(Return)

Function(Function)

Implementations

impl Stmt

pub fn new_raw(s: &str) -> Self

pub fn new_non_printable(s: &str) -> Self

pub fn new_print(expr: Expr) -> Self

pub fn new_comment(s: &str) -> Self

pub fn new_assign(var: Variable, expr: Expr) -> Self

pub fn new_return(expr: Option<Expr>) -> Self