Struct osmia::Osmia

pub struct Osmia;

This simple API is designed to use the osmia template engine.

Syntax

Text:

By default, all text outside the delimiters will be treated as a plain text.

Hey there! This is a plain text

Hey there! This is a plain text

Expression:

Evaluates and prints the given expression.

1 + 2 = {{ 1 + 2 }}

1 + 2 = 3

Assign:

Assigns the given expression to the given variable in the context.

{{ assign foo = 1 + 2 }}
The result is {{foo}}

The result is 3

Print:

Prints the given expression in the stdin. It can be useful when debugging.

{{ print "Hello world!" }}

If:

Single if:

{{ if 1 == 1 }}
Math is not broken!
{{ fi }}

Math is not broken!

If, else:

{{ if 1 == 2 }}
Math is not broken!
{{ else }}
Nope, math is right >D
{{ fi }}

Nope, math is right >D

If, else if:

{{assign value = 2}}
{{ if value == 1 }}
Nope
{{ elseif value == 2 }}
The value is 2!
{{ else }}
Nope
{{ fi }}

The value is 2!

While:

{{ assign i = 0 }}
{{ while i < 5 }}
{{ assign i = i + 1 }}
i = {{ i }}
{{done}}

i = 1
i = 2
i = 3
i = 4
i = 5

For:

{{ for i in [1, 2, 3] }}
  i = {{ i }}
{{ done }}
{{assign arr = [
  {"name": "Marvin1"},
  {"name": "Marvin2"},
  {"name": "Marvin3"}
]}}
{{for user in arr}}
User: {{ user.name }}
{{done}}

i = 1
i = 2
i = 3
User: Marvin1
User: Marvin2
User: Marvin3

Continue, break:

{{continue}}
{{break}}

Advanced control:

All variable in the context can be used and edited similarly to variables in JavaScript.

{{ assign obj = {"user": {"name": "Marvin"} } }}
Old name: {{ obj.user.name }}
{{ assign obj.user.name = "R2D2" }}
New name: {{ obj.user.name }}

Old name: Marvin
New name: R2D2

Examples

Basic execution

use osmia::Osmia;

let mut interpreter = Osmia::new();
let code = Osmia::code("Hello world! {{ 1 + 2 }}").unwrap();
let result = interpreter.run(&code).unwrap();
assert_eq!(result, "Hello world! 3");

Context

The context can be passed as a JSON object in the form of a string.

use osmia::Osmia;

let context_json = r#"{"foo": "bar"}"#;
let code = "The value of foo is {{ foo }}";
let mut interpreter = Osmia::from_json(context_json).unwrap();
let code = Osmia::code(code).unwrap();
let result = interpreter.run(&code).unwrap();
assert_eq!(result, "The value of foo is bar");

Experimental features: Custom delimiters

You can change the delimiters used by the code lexer.

use osmia::Osmia;

let mut interpreter = Osmia::from_json(r#"{"user": {"name": "Marvin"}}"#).unwrap();
let code = Osmia::custom_code("User: ${ user.name }", "${", "}").unwrap();
let result = interpreter.run(&code).unwrap();
assert_eq!(result, "User: Marvin");

Implementations

impl Osmia

pub const VERSION: &'static str = "1.13.2"

pub fn new() -> Interpreter

pub fn from_json(json: &str) -> Result<Interpreter, String>

pub fn code(code: &str) -> Result<Stmt, String>

pub fn custom_code( code: &str, start_delimiter: &str, end_delimiter: &str ) -> Result<Stmt, String>