osmia::model::ctx::json_tree_error

Enum JsonTreeError

pub enum JsonTreeError<K> {
    AccessValue(K),
    ArrayOutOfBounds((usize, usize)),
    IndexInObject,
    KeyInArray,
    KeyNotFound(K),
    NoKey,
}

Collection of errors that can occur while traversing a JsonTree. When the error contains any value, it represents the error location.

Variants

AccessValue(K)

Attempted to access a value.

Examples:

{"foo": "bar"}

foo.something_not_valid
foo[1]

ArrayOutOfBounds((usize, usize))

Attempted access an array with an index that is out of bounds. By default, the first element is the index attempted while the second is the size of the array.

Examples:

{"foo": ["bar"]}

foo[1]
foo[123]

IndexInObject

Attempted to access an object with an index.

Examples:

{"foo": {"bar": "baz"}}

foo[1]
foo.bar[0]

KeyInArray

Attempted to access an array with a key.

Example:

{"foo": ["bar"]}

foo.something_not_valid

KeyNotFound(K)

Attempted to access an object with a key that does not exist.

Example:

{"foo": {"bar": "baz"}}

foo.something_not_valid

NoKey

When no key has been provided. Note: In practice, this should never happen.

Example:

Imagine we have:

{"foo": "bar"}

And we attempt to assign a new value to the whole object. In that case, this error will be returned.

Implementations

impl JsonTreeError<JsonTreeKey<String>>

fn format_get_error(self) -> String

fn format_set_error(self) -> String